+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
-- test: testing release automation after sonatype migration
+## 3.18.0 - 2025-06-12
+
 - feat: add proxy to `PostHogConfig` ([#260](https://github.com/PostHog/posthog-android/issues/260))
     - Thanks @MamboBryan